Fix a debug message improperly saying that we were unable to find a properties file. PCI.properties updated to 2; will now attempt to replace any bad references to old class names. (aka will attempt to replace: com.shanebeestudios.api to com.itsschatten.libs) Update Furnace API to 1.1.4
Remove Permission#EMPTY, unused and never really worked in the first place.
Minor refactoring to the SignListener#getSign() method; removes some duplicate code.
Update SignListener to use Java 16 features.
As there are no updates to any of the modules that handle NMS code the versions of those modules is not updated.